Extract cluster inspector from manager
